requirements.txt
streamlit
pytesseract
pillow
googletrans==4.0.0-rc1
pdf2image
python-docx